fix: Java enum constants links were broken.
Doxygen entities of "kind" equal to "variable" with empty `<type>` child tag, which are mapped to internal  Kind::JAVAENUMCONSTANT (Java enums constants) were written as "enum constant" (with space) in the generated links URLs. It causes links to be broken. Spaces must be replaced with `-` instead of space, this is what Markdown processors use to replace spaces with in generated anchors.