feat: Fix Java links being inside code blocks.
When processing links in Java code (`{@link class#method(argType1, argType2, argType3)}`) Doxygen wraps them first in the `<computeroutput>` tag (a Doxygen semantic for `<code>` tag in Markdown) and inside in the `<ref>` tag (used for creating various links).

It leads to putting links inside code, which is not valid, those links will not be parsed, unless `linkAndInlineCodeAsHTML` config option is used (which can be also problematic in some cases).

Added a hack-tweak that when we encounter such case (a `<computeroutput>` tag with a single `<ref>` tag in it), and `linkAndInlineCodeAsHTML` option is not set, we effectively inverting them, by putting links braces over code marks (backtick).